Move datasource redeem flag to NFT section
