+++ src/components/Create/components/pages/ReconfigurationRules/hooks/ReconfigurationRulesForm.ts
@@ -16,7 +16,6 @@ type ReconfigurationRulesFormProps = Partial<{
   pausePayments: boolean
   allowTerminalConfiguration: boolean
   holdFees: boolean
-  useDataSourceForRedeem: boolean
 }>
 
 export const useReconfigurationRulesForm = () => {
@@ -134,14 +133,6 @@ export const useReconfigurationRulesForm = () => {
     fieldName: 'holdFees',
     ignoreUndefined: true,
     dispatchFunction: editingV2ProjectActions.setHoldFees,
-    formatter: v => !!v,
-  })
-
-  useFormDispatchWatch({
-    form,
-    fieldName: 'useDataSourceForRedeem',
-    ignoreUndefined: true,
-    dispatchFunction: editingV2ProjectActions.setUseDataSourceForRedeem,
     formatter: v => !!v,
   })
