+++ src/components/Create/components/pages/ReviewDeploy/components/RewardsReview/RewardsReview.tsx
@@ -1,14 +1,22 @@
+import { t } from '@lingui/macro'
+import { Row } from 'antd'
 import { Reward, RewardsList } from 'components/Create/components/RewardsList'
 import { useAppDispatch } from 'hooks/AppDispatch'
 import { useAppSelector } from 'hooks/AppSelector'
 import { useCallback, useMemo } from 'react'
 import { editingV2ProjectActions } from 'redux/slices/editingV2Project'
+import { formatEnabled } from 'utils/format/formatBoolean'
 import { v4 } from 'uuid'
+import { DescriptionCol } from '../DescriptionCol'
 
 export const RewardsReview = () => {
   const { rewardTiers } = useAppSelector(
     state => state.editingV2Project.nftRewards,
   )
+  const fundingCycleMetadata = useAppSelector(
+    state => state.editingV2Project.fundingCycleMetadata,
+  )
+
   const dispatch = useAppDispatch()
 
   const rewards = useMemo(() => {
@@ -50,5 +58,24 @@ export const RewardsReview = () => {
     [dispatch],
   )
 
-  return <RewardsList value={rewards} onChange={setRewards} />
+  const useDataSourceForRedeem = useMemo(() => {
+    return formatEnabled(fundingCycleMetadata.useDataSourceForRedeem)
+  }, [fundingCycleMetadata.useDataSourceForRedeem])
+
+  return (
+    <>
+      <RewardsList value={rewards} onChange={setRewards} />
+      <Row gutter={20}>
+        <DescriptionCol
+          span={24}
+          title={t`Redeemable NFTs`}
+          desc={
+            <div className="text-base font-medium">
+              {useDataSourceForRedeem}
+            </div>
+          }
+        />
+      </Row>
+    </>
+  )
 }
