+++ src/components/Create/components/pages/ReviewDeploy/components/RulesReview/MobileRulesReview.tsx
@@ -11,8 +11,8 @@ export const MobileRulesReview = () => {
     strategy,
     terminalConfiguration,
     holdFees,
-    useDataSourceForRedeem,
   } = useRulesReview()
+
   return (
     <Row gutter={[20, 20]}>
       <DescriptionCol
@@ -46,13 +46,6 @@ export const MobileRulesReview = () => {
         span={12}
         title={t`Hold fees`}
         desc={<div className="text-base font-medium">{holdFees}</div>}
-      />
-      <DescriptionCol
-        span={12}
-        title={t`Use data source for redeem`}
-        desc={
-          <div className="text-base font-medium">{useDataSourceForRedeem}</div>
-        }
       />
     </Row>
   )
